Rename connect actions to match other actions
