+++ src/react/Actions/share_invite_bank_responses.js
