+++ src/react/Pages/AccountInfo.jsx
@@ -40,8 +40,8 @@ import { requestResponsesUpdate } from "../Actions/request_responses";
 import { bunqMeTabsUpdate } from "../Actions/bunq_me_tabs";
 import { masterCardActionsUpdate } from "../Actions/master_card_actions";
 import { requestInquiriesUpdate } from "../Actions/request_inquiries";
-import { shareInviteBankInquiriesInfoUpdate } from "../Actions/share_invite_bank_inquiry";
-import { shareInviteBankResponsesInfoUpdate } from "../Actions/share_invite_bank_response";
+import { shareInviteBankInquiriesInfoUpdate } from "../Actions/share_invite_bank_inquiries";
+import { shareInviteBankResponsesInfoUpdate } from "../Actions/share_invite_bank_responses";
 
 const styles = {
     textField: {
